Add gizmo for adding/removing tags
The current interface has an issue where it takes quite a few clicks to manage the tags for a group of pawns all at once, especially if you want to pre-select them based preset criteria. For example, if I have 50 colonists and want to assign them tags based on their shooting ability, I would need to predetermine which pawns I want to assign, take a note in  another application, and then go to the tag window to use the mass assignment there. This commit introduces a simple set of gizmos that allows a user to apply or remove tags from all pawns currently selected.