improve 'Better Pawn Control' support
Binding only the panic state to the alert button of BPC
leads to an inconsistent experience - the primary functionality of BPC
is to switch policies and the alert button merely enables/reverts a specific
policy. It should be the same as activating the relevant policies manually,
but that did not work.

This patch makes BPC fully responsible for setting the loadout state when
the alert button is activated, and also makes BPC handle properly set loadout
states when BPC policy is changed. Requires support in BPC (submitted there).